// src/services/models/response-result-get-company-danger-factor-and-solution-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { GetCompanyDangerFactorAndSolutionResponse } from './get-company-danger-factor-and-solution-response';

/**
 * 
 * @export
 * @interface ResponseResultGetCompanyDangerFactorAndSolutionResponse
 */
export interface ResponseResultGetCompanyDangerFactorAndSolutionResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyDangerFactorAndSolutionResponse
     */
    'code'?: string;
    /**
     * 
     * @type {GetCompanyDangerFactorAndSolutionResponse}
     * @memberof ResponseResultGetCompanyDangerFactorAndSolutionResponse
     */
    'data'?: GetCompanyDangerFactorAndSolutionResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyDangerFactorAndSolutionResponse
     */
    'message'?: string;
}